refactor: modularize entities, rendering, utils, and docstrings, restructured directories, moved to poetry (0023)
