+++ .archive/chat-transitions/transition-004.md
@@ -1,0 +1,79 @@
+# chat-transition/transition-003.md
+
+## Transition Layer 24
+
+**Purpose:**  
+Transition layer to document the migration to a modular codebase and updated project workflow. No features or code changes are included in this layer.
+
+---
+
+### Summary of This Transition
+
+- **Modular refactor completed in Layer 23:**
+  - All entity logic moved to `entities.py`
+  - Rendering logic moved to `renderer.py`
+  - Utilities and shared constants moved to `utils.py`
+  - `game.py` now orchestrates all major components
+- **Project rules clarified and formalized** (see README.md and rules section below)
+- **Python installation and environment setup instructions moved to [PYTHON_SETUP.md](../PYTHON_SETUP.md)**
+- **Directory structure and all documentation updated for clarity and traceability**
+- **Prompt index (`PROMPT_INDEX.md`) will be updated only after merge to main, not during feature or transition work**
+
+---
+
+### New and Reinforced Project Rules
+
+- Every new feature, fix, or refactor is developed as a “layer,” beginning with a prompt and ending with `commit layer`
+- `test`: Generate and display full, ready-to-run code for all affected files in a single markdown block
+- `preview`: Show updated code or artifact preview after any change during the test cycle
+- `complete`: Lock in the current tested code as the layer output, generate all logs/artifacts, and prepare the commit
+- Every devlog must include the full, verbatim transcript of the conversation for that layer
+- All documentation and code previews must be shown as a single, contiguous markdown block—never split across multiple chat messages
+- Prompt index is only updated after merge to main (to avoid hash/order conflicts)
+- All commit messages must end with the layer number in parentheses, e.g., `(0024)`
+- All code and documentation must be LLM-generated and traceable to a prompt and commit
+- Emoji are never used in code, documentation, or logs
+
+---
+
+### Updated Directory Structure
+
+game-example/
+├── src/
+│ ├── game.py # Main entry point, game loop, orchestration
+│ ├── entities.py # Player, Adversary, and all entity logic/state
+│ ├── renderer.py # Rendering, colorization, and display logic
+│ ├── utils.py # Utility functions, terminal helpers, shared constants
+│ └── mapgen.py # Map generation logic
+│
+├── debug/
+│ └── debug-XXXX-a/
+│ ├── a1/
+│ ├── a2/
+│ └── README.md
+│
+├── chat-transition/
+│ ├── transition-001.md
+│ ├── transition-002.md
+│ └── transition-003.md 
+| └── transition-004.md *this file*
+│
+├── PROMPTS.md
+├── CHAT_LOG.md
+├── PROMPT_INDEX.md
+├── AUTHOR_NOTES.md
+├── README.md
+└── PYTHON_SETUP.md
+
+
+---
+
+### Status
+
+- **Current layer:** 24 (transition)
+- **Last feature layer:** 23 (full modularization)
+- **Prompt index** to be updated after next merge to main.
+
+---
+
+**This file documents all structural and policy changes for this session.**
