+++ bin/play_game.sh
@@ -1,0 +1,1 @@
+ python app/src/game.py
