+++ logs/debug/debug-0012-a/a2/mapgen.py
